pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperspeed-nostalgia"
version = "0.1.0"
description = "Reclaiming AI artifacts through art historical parallels"
readme = "HYPERSPEED_NOSTALGIA_Project_Document_v1.2.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Human + Claude collaboration"}
]
keywords = ["ai", "art", "deepdream", "stable-diffusion", "artifacts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Artistic Software",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
hyperspeed = "hyperspeed.cli:app"

[project.urls]
Homepage = "https://github.com/example/hyperspeed-nostalgia"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
